+++ sdk/azcore/CHANGELOG.md
@@ -57,6 +57,8 @@
 ### Bugs Fixed
 
 * Suppress creating spans for nested SDK API calls. The HTTP span will be a child of the outer API span.
+* Fix default HTTP transport to work in WASM modules.
+* Enable HTTP connection health checks for HTTP/2 connections.
 
 ### Other Changes